Flyttet ut vurderinger i egne filer
