--- src/main/kotlin/no/nav/tilleggsstonader/sak/vilkår/vilkårperiode/domain/faktavurderinger/TypeFaktaOgVurdering.kt
@@ -1,7 +1,0 @@
-package no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.faktavurderinger
-
-import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.VilkårperiodeType
-
-sealed interface TypeFaktaOgVurdering {
-    val vilkårperiodeType: VilkårperiodeType
-}
