+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vilkår/vilkårperiode/domain/faktavurderinger/VurderingerMålgruppe.kt
@@ -1,0 +1,37 @@
+package no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.faktavurderinger
+
+import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.DelvilkårVilkårperiode
+
+sealed interface MedlemskapVurdering : Vurderinger {
+    val medlemskap: DelvilkårVilkårperiode.Vurdering
+}
+
+sealed interface DekketAvAnnetRegelverkVurdering : Vurderinger {
+    val dekketAvAnnetRegelverk: DelvilkårVilkårperiode.Vurdering
+}
+
+data class VurderingAAP(
+    override val dekketAvAnnetRegelverk: DelvilkårVilkårperiode.Vurdering,
+) : MedlemskapVurdering, DekketAvAnnetRegelverkVurdering {
+    override val medlemskap: DelvilkårVilkårperiode.Vurdering =
+        DelvilkårVilkårperiode.Vurdering.VURDERING_IMPLISITT_OPPFYLT
+}
+
+data class VurderingUføretrygd(
+    override val medlemskap: DelvilkårVilkårperiode.Vurdering,
+    override val dekketAvAnnetRegelverk: DelvilkårVilkårperiode.Vurdering,
+) : MedlemskapVurdering, DekketAvAnnetRegelverkVurdering
+
+data class VurderingNedsattArbeidsevne(
+    override val medlemskap: DelvilkårVilkårperiode.Vurdering,
+    override val dekketAvAnnetRegelverk: DelvilkårVilkårperiode.Vurdering,
+) : MedlemskapVurdering, DekketAvAnnetRegelverkVurdering
+
+data class VurderingOmstillingsstønad(
+    override val medlemskap: DelvilkårVilkårperiode.Vurdering,
+) : MedlemskapVurdering
+
+data object VurderingOvergangsstønad : MedlemskapVurdering {
+    override val medlemskap: DelvilkårVilkårperiode.Vurdering =
+        DelvilkårVilkårperiode.Vurdering.VURDERING_IMPLISITT_OPPFYLT
+}
